+++ src/combinator.rs
@@ -118,13 +118,13 @@ impl<T> Combinator<T> {
   }
 }
 
-unsafe impl<Text: ?Sized, State, Heap, T: Action<Text, State, Heap>> Action<Text, State, Heap>
+unsafe impl<TextRef, State, Heap, T: Action<TextRef, State, Heap>> Action<TextRef, State, Heap>
   for Combinator<T>
 {
   type Value = T::Value;
 
   #[inline]
-  fn exec(&self, input: Input<&Text, &mut State, &mut Heap>) -> Option<Output<T::Value>> {
+  fn exec(&self, input: Input<TextRef, &mut State, &mut Heap>) -> Option<Output<T::Value>> {
     self.action.exec(input)
   }
 }