+++ src/combinator/decorator/flow.rs
@@ -10,17 +10,17 @@ create_simple_decorator!(Optional, "See [`Combinator::optional`].");
 create_simple_decorator!(Boundary, "See [`Combinator::boundary`].");
 
 unsafe impl<
-    Text: ?Sized,
+    TextRef: Clone,
     State,
     Heap,
-    T: Action<Text, State, Heap>,
-    D: Fn(Input<&Text, &mut State, &mut Heap>) -> bool,
-  > Action<Text, State, Heap> for When<T, D>
+    T: Action<TextRef, State, Heap>,
+    D: Fn(Input<TextRef, &mut State, &mut Heap>) -> bool,
+  > Action<TextRef, State, Heap> for When<T, D>
 {
   type Value = T::Value;
 
   #[inline]
-  fn exec(&self, mut input: Input<&Text, &mut State, &mut Heap>) -> Option<Output<Self::Value>> {
+  fn exec(&self, mut input: Input<TextRef, &mut State, &mut Heap>) -> Option<Output<Self::Value>> {
     if (self.inner)(input.reborrow()) {
       self.action.exec(input)
     } else {
@@ -30,17 +30,17 @@ unsafe impl<
 }
 
 unsafe impl<
-    Text: ?Sized,
+    TextRef: Clone,
     State,
     Heap,
-    T: Action<Text, State, Heap>,
-    D: Fn(Input<&Text, &mut State, &mut Heap>) -> bool,
-  > Action<Text, State, Heap> for Prevent<T, D>
+    T: Action<TextRef, State, Heap>,
+    D: Fn(Input<TextRef, &mut State, &mut Heap>) -> bool,
+  > Action<TextRef, State, Heap> for Prevent<T, D>
 {
   type Value = T::Value;
 
   #[inline]
-  fn exec(&self, mut input: Input<&Text, &mut State, &mut Heap>) -> Option<Output<Self::Value>> {
+  fn exec(&self, mut input: Input<TextRef, &mut State, &mut Heap>) -> Option<Output<Self::Value>> {
     if !(self.inner)(input.reborrow()) {
       self.action.exec(input)
     } else {
@@ -50,17 +50,17 @@ unsafe impl<
 }
 
 unsafe impl<
-    Text: ?Sized,
+    TextRef: Clone,
     State,
     Heap,
-    T: Action<Text, State, Heap>,
-    D: Fn(AcceptedContext<Input<&Text, &mut State, &mut Heap>, &Output<T::Value>>) -> bool,
-  > Action<Text, State, Heap> for Reject<T, D>
+    T: Action<TextRef, State, Heap>,
+    D: Fn(AcceptedContext<Input<TextRef, &mut State, &mut Heap>, &Output<T::Value>>) -> bool,
+  > Action<TextRef, State, Heap> for Reject<T, D>
 {
   type Value = T::Value;
 
   #[inline]
-  fn exec(&self, mut input: Input<&Text, &mut State, &mut Heap>) -> Option<Output<Self::Value>> {
+  fn exec(&self, mut input: Input<TextRef, &mut State, &mut Heap>) -> Option<Output<Self::Value>> {
     self.action.exec(input.reborrow()).and_then(|output| {
       if (self.inner)(AcceptedContext {
         input,
@@ -74,13 +74,13 @@ unsafe impl<
   }
 }
 
-unsafe impl<Text: ?Sized, State, Heap, T: Action<Text, State, Heap, Value: Default>>
-  Action<Text, State, Heap> for Optional<T>
+unsafe impl<TextRef, State, Heap, T: Action<TextRef, State, Heap, Value: Default>>
+  Action<TextRef, State, Heap> for Optional<T>
 {
   type Value = T::Value;
 
   #[inline]
-  fn exec(&self, input: Input<&Text, &mut State, &mut Heap>) -> Option<Output<Self::Value>> {
+  fn exec(&self, input: Input<TextRef, &mut State, &mut Heap>) -> Option<Output<Self::Value>> {
     Some(self.action.exec(input).unwrap_or_else(|| Output {
       value: Default::default(),
       digested: 0,
@@ -88,11 +88,13 @@ unsafe impl<Text: ?Sized, State, Heap, T: Action<Text, State, Heap, Value: Defau
   }
 }
 
-unsafe impl<State, Heap, T: Action<str, State, Heap>> Action<str, State, Heap> for Boundary<T> {
+unsafe impl<'a, State, Heap, T: Action<&'a str, State, Heap>> Action<&'a str, State, Heap>
+  for Boundary<T>
+{
   type Value = T::Value;
 
   #[inline]
-  fn exec(&self, mut input: Input<&str, &mut State, &mut Heap>) -> Option<Output<Self::Value>> {
+  fn exec(&self, mut input: Input<&'a str, &mut State, &mut Heap>) -> Option<Output<Self::Value>> {
     self.action.exec(input.reborrow()).and_then(|output| {
       unsafe { input.instant().rest().get_unchecked(output.digested..) }
         .chars()
@@ -117,12 +119,12 @@ impl<T> Combinator<T> {
   /// # ;}
   /// ```
   #[inline]
-  pub fn when<Text: ?Sized, State, Heap, F: Fn(Input<&Text, &mut State, &mut Heap>) -> bool>(
+  pub fn when<TextRef, State, Heap, F: Fn(Input<TextRef, &mut State, &mut Heap>) -> bool>(
     self,
     condition: F,
   ) -> Combinator<When<T, F>>
   where
-    T: Action<Text, State, Heap>,
+    T: Action<TextRef, State, Heap>,
   {
     Combinator::new(When::new(self.action, condition))
   }
@@ -140,12 +142,12 @@ impl<T> Combinator<T> {
   /// # ;}
   /// ```
   #[inline]
-  pub fn prevent<Text: ?Sized, State, Heap, F: Fn(Input<&Text, &mut State, &mut Heap>) -> bool>(
+  pub fn prevent<TextRef, State, Heap, F: Fn(Input<TextRef, &mut State, &mut Heap>) -> bool>(
     self,
     preventer: F,
   ) -> Combinator<Prevent<T, F>>
   where
-    T: Action<Text, State, Heap>,
+    T: Action<TextRef, State, Heap>,
   {
     Combinator::new(Prevent::new(self.action, preventer))
   }
@@ -161,16 +163,16 @@ impl<T> Combinator<T> {
   /// ```
   #[inline]
   pub fn reject<
-    Text: ?Sized,
+    TextRef,
     State,
     Heap,
-    F: Fn(AcceptedContext<Input<&Text, &mut State, &mut Heap>, &Output<T::Value>>) -> bool,
+    F: Fn(AcceptedContext<Input<TextRef, &mut State, &mut Heap>, &Output<T::Value>>) -> bool,
   >(
     self,
     rejecter: F,
   ) -> Combinator<Reject<T, F>>
   where
-    T: Action<Text, State, Heap>,
+    T: Action<TextRef, State, Heap>,
   {
     Combinator::new(Reject::new(self.action, rejecter))
   }
@@ -236,14 +238,14 @@ mod tests {
   use super::*;
   use crate::{combinator::wrap, instant::Instant};
 
-  fn accepter() -> Combinator<impl Action<str, bool, Value = ()>> {
+  fn accepter() -> Combinator<impl for<'a> Action<&'a str, bool, Value = ()>> {
     wrap(|input| {
       *input.state = true;
       input.digest(1)
     })
   }
 
-  fn rejecter() -> Combinator<impl Action<str, bool, Value = ()>> {
+  fn rejecter() -> Combinator<impl for<'a> Action<&'a str, bool, Value = ()>> {
     wrap(|input| {
       *input.state = true;
       None