+++ src/combinator/ops/bitor.rs
@@ -49,17 +49,17 @@ impl<Lhs, Rhs> BitOr<Lhs, Rhs> {
 }
 
 unsafe impl<
-    Text: ?Sized,
+    TextRef: Clone,
     State,
     Heap,
-    Lhs: Action<Text, State, Heap>,
-    Rhs: Action<Text, State, Heap, Value = Lhs::Value>,
-  > Action<Text, State, Heap> for BitOr<Lhs, Rhs>
+    Lhs: Action<TextRef, State, Heap>,
+    Rhs: Action<TextRef, State, Heap, Value = Lhs::Value>,
+  > Action<TextRef, State, Heap> for BitOr<Lhs, Rhs>
 {
   type Value = Lhs::Value;
 
   #[inline]
-  fn exec(&self, mut input: Input<&Text, &mut State, &mut Heap>) -> Option<Output<Self::Value>> {
+  fn exec(&self, mut input: Input<TextRef, &mut State, &mut Heap>) -> Option<Output<Self::Value>> {
     self
       .lhs
       .exec(input.reborrow())