+++ src/combinator/ops/mul/inline.rs
@@ -51,19 +51,19 @@ impl<T> Combinator<T> {
   /// ```
   #[inline]
   pub fn fold<
-    Text: ?Sized,
+    TextRef,
     State,
     Heap,
     Acc,
     Init: Fn() -> Acc,
-    Folder: Fn(T::Value, Acc, Input<&Text, &mut State, &mut Heap>) -> Acc,
+    Folder: Fn(T::Value, Acc, Input<TextRef, &mut State, &mut Heap>) -> Acc,
   >(
     self,
     init: Init,
     folder: Folder,
   ) -> InlineFold<T, Init, Folder>
   where
-    T: Action<Text, State, Heap>,
+    T: Action<TextRef, State, Heap>,
   {
     InlineFold {
       action: self.action,
@@ -84,22 +84,20 @@ impl<T, Init, Folder, Repeater: Repeat> ops::Mul<Repeater> for InlineFold<T, Ini
 }
 
 unsafe impl<
-    Text: ?Sized,
+    TextRef: Digest + Clone,
     State,
     Heap,
-    T: Action<Text, State, Heap>,
+    T: Action<TextRef, State, Heap>,
     Acc,
     Repeater: Repeat,
     Init: Fn() -> Acc,
-    Folder: Fn(T::Value, Acc, Input<&Text, &mut State, &mut Heap>) -> Acc,
-  > Action<Text, State, Heap> for Mul<InlineFold<T, Init, Folder>, Repeater>
-where
-  for<'a> &'a Text: Digest,
+    Folder: Fn(T::Value, Acc, Input<TextRef, &mut State, &mut Heap>) -> Acc,
+  > Action<TextRef, State, Heap> for Mul<InlineFold<T, Init, Folder>, Repeater>
 {
   type Value = Acc;
 
   #[inline]
-  fn exec(&self, mut input: Input<&Text, &mut State, &mut Heap>) -> Option<Output<Self::Value>> {
+  fn exec(&self, mut input: Input<TextRef, &mut State, &mut Heap>) -> Option<Output<Self::Value>> {
     let repeat = &self.rhs;
     impl_mul!(input, repeat, self.lhs.init, self.lhs.fold, self.lhs.action)
   }