+++ CHANGELOG.md
@@ -18,7 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Input` now take an additional generic param `TextRef` instead of a lifetime param.
 - The `rest` of `Input::instant` is no longer guaranteed to be non-empty.
 - `Input::new` will return `Input` directly instead of an `Option`.
-- `Action` now take an additional generic param `Text`.
+- `Action` now take an additional generic param `TextRef`.
+- `Parser` now take an additional generic param `TextRef` instead of a lifetime param.
 - `Snapshot` now take an additional generic param `TextRef` instead of a lifetime param.
 
 ### Fixed